hirbをpryで用いられるように設定 保存時のvalidation組み込み中 列除去・追加を条件分岐のマイグレーションで設定

+++ db/migrate/20161010075746_re_create_fields_to_user_for_autocomplete.rb
@@ -1,16 +1,21 @@
 class ReCreateFieldsToUserForAutocomplete < ActiveRecord::Migration[5.0]
   def change
-    remove_column :users, :first_name 
-    remove_column :users, :last_name 
-    remove_column :users, :address 
-    remove_column :users, :post_code
+    if column_exists? :users, :first_name
+      remove_column :users, :first_name 
+      remove_column :users, :last_name 
+      remove_column :users, :address 
+      remove_column :users, :post_code
+    end
 
-    add_column :users, :family_name, :string 
-    add_column :users, :given_name, :string 
-    add_column :users, :postal_code, :string
-    add_column :users, :address_level1, :string 
-    add_column :users, :address_level2, :string
-    add_column :users, :address_line1, :string 
-    add_column :users, :address_line2, :string 
+    unless column_exists? :users, :family_name
+      add_column :users, :family_name, :string  
+      add_column :users, :given_name, :string 
+      add_column :users, :postal_code, :string
+      add_column :users, :address_level1, :string 
+      add_column :users, :address_level2, :string
+      add_column :users, :address_line1, :string 
+      add_column :users, :address_line2, :string
+    end
+     
   end
 end
